requirements.txt
asgiref==3.5.0
Django==2.2
djangorestframework==3.12.4
pytz==2022.1
sqlparse==0.4.2
typing-extensions==4.1.1
